+++ packages/code/src/Syntax.tsx
@@ -11,6 +11,7 @@ import { SupportedLanguages, languageParsers } from './languages';
 import { injectGlobalStyles } from './globalStyles';
 import renderingPlugin, { TableContent } from './renderingPlugin';
 import { SyntaxContext } from './SyntaxContext';
+import { useBaseFontSize } from '@leafygreen-ui/leafygreen-provider';
 
 type FilteredSupportedLanguagesEnum = Omit<
   typeof SupportedLanguages,
@@ -98,6 +99,14 @@ function Syntax({
 
   const mode = darkMode ? Mode.Dark : Mode.Light;
 
+  const baseFontSize = useBaseFontSize();
+  // TODO: Refresh - remove darkMode logic
+  const codeFontSize = baseFontSize === 14 ? 13 : darkMode ? 16 : 15;
+
+  const codeFontStyles = css`
+    font-size: ${codeFontSize}px;
+  `;
+
   return (
     <SyntaxContext.Provider
       value={{
@@ -112,6 +121,7 @@ function Syntax({
         className={cx(
           `lg-highlight-hljs-${mode}`,
           codeStyles,
+          codeFontStyles,
           language,
           className,
         )}