+++ packages/modal/src/index.ts
@@ -1,5 +1,5 @@
 import Footer from './Footer';
-import Modal, { ModalSize } from './Modal';
+import Modal, { ModalSize, CloseIconColor } from './Modal';
 
-export { Footer, ModalSize };
+export { Footer, ModalSize, CloseIconColor };
 export default Modal;