+++ packages/tooltip/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @leafygreen-ui/tooltip
 
+## 7.0.0
+
+### Major Changes
+
+- ab1fd9e: - Updates tooltip for visual brand refresh.
+  - Updates darkMode for Tooltip
+  - Using `usePortal={false}` will set the tooltip to the width of the content using `max-content` (up to a `max-width` of 256px). You can override this behavior using `portalClassName`. (Override the max-width using `className` on the Tooltip)
+
 ## 6.3.1
 
 ### Patch Changes